refactor(action): apply ponytail cuts to Node pipeline
Collapse five modules into lib.mjs, drop DI/typedef/PATH ceremony, and
simplify Windows install/PATH steps without changing the public action API.
