+++ docs/src/content/docs/ci-cd/github-action.mdx
@@ -7,8 +7,6 @@ import { Aside } from '@astrojs/starlight/components';
 
 Vizb provides a composite GitHub Action to turn CSV/JSON tabular data or benchmark output (Go, Rust, JavaScript) into interactive HTML visualizations in CI. Add it to any workflow with a single step.
 
-Install and cache steps stay in the composite YAML; convert / merge / UI orchestration runs as a zero-dependency Node script shipped with the action.
-
 ## Basic Usage
 
 ```yaml
